@@ -1,12 +1,20 @@
 import React from 'react'
-import { Route, Switch, Redirect } from 'react-router-dom'
+import {
+  Route,
+  Routes as RouterRoutes,
+  Navigate,
+  useNavigate,
+  NavigateFunction,
+  Outlet,
+} from 'react-router-dom'
 
 import Layout from '../layout/Layout'
-import { clearTokenCookie } from '../../helpers/authentication'
-import { useTranslation } from 'react-i18next'
+import { authToken, clearTokenCookie } from '../../helpers/authentication'
+import { TFunction, useTranslation } from 'react-i18next'
 import Loader from '../../primitives/Loader'
-
-const AuthorizedRoute = React.lazy(() => import('./AuthorizedRoute'))
+import AuthorizedRoute from './AuthorizedRoute'
+import sharePageRoute from '../../Pages/SharePage/sharePageRoute'
+import peoplePageRoute from '../../Pages/PeoplePage/peoplePageRoute'
 
 const AlbumsPage = React.lazy(
   () => import('../../Pages/AllAlbumsPage/AlbumsPage')
@@ -16,8 +24,6 @@ const TimelinePage = React.lazy(
   () => import('../../Pages/TimelinePage/TimelinePage')
 )
 const PlacesPage = React.lazy(() => import('../../Pages/PlacesPage/PlacesPage'))
-const SharePage = React.lazy(() => import('../../Pages/SharePage/SharePage'))
-const PeoplePage = React.lazy(() => import('../../Pages/PeoplePage/PeoplePage'))
 
 const LoginPage = React.lazy(() => import('../../Pages/LoginPage/LoginPage'))
 const InitialSetupPage = React.lazy(
@@ -30,6 +36,7 @@ const SettingsPage = React.lazy(
 
 const Routes = () => {
   const { t } = useTranslation()
+  const navigate = useNavigate()
 
   return (
     <React.Suspense
@@ -39,37 +46,86 @@ const Routes = () => {
         </Layout>
       }
     >
-      <Switch>
-        <Route path="/login" component={LoginPage} />
-        <Route path="/logout">
-          {() => {
-            clearTokenCookie()
-            location.href = '/'
-          }}
+      <RouterRoutes>
+        <Route path="/login" element={<LoginPage />} />
+        <Route path="/logout" element={<LogoutPage navigate={navigate} />} />
+        <Route path="/initialSetup" element={<InitialSetupPage />} />
+        <Route path="/share">{sharePageRoute({ t })}</Route>
+        <Route
+          path="/albums"
+          element={
+            <AuthorizedRoute>
+              <AlbumsPage />
+            </AuthorizedRoute>
+          }
+        />
+        <Route
+          path="/album/:id"
+          element={
+            <AuthorizedRoute>
+              <AlbumPage />
+            </AuthorizedRoute>
+          }
+        />
+        <Route
+          path="/timeline"
+          element={
+            <AuthorizedRoute>
+              <TimelinePage />
+            </AuthorizedRoute>
+          }
+        />
+        <Route
+          path="/places"
+          element={
+            <AuthorizedRoute>
+              <PlacesPage />
+            </AuthorizedRoute>
+          }
+        />
+        <Route
+          path="/people"
+          element={
+            <AuthorizedRoute>
+              <Outlet />
+            </AuthorizedRoute>
+          }
+        >
+          {peoplePageRoute()}
         </Route>
-        <Route path="/initialSetup" component={InitialSetupPage} />
-        <Route path="/share" component={SharePage} />
-        <AuthorizedRoute exact path="/albums" component={AlbumsPage} />
-        <AuthorizedRoute path="/album/:id" component={AlbumPage} />
-        <AuthorizedRoute path="/timeline" component={TimelinePage} />
-        <AuthorizedRoute path="/places" component={PlacesPage} />
-        <AuthorizedRoute path="/people/:person?" component={PeoplePage} />
-        <AuthorizedRoute path="/settings" component={SettingsPage} />
-        <Route path="/" exact render={() => <Redirect to="/timeline" />} />
+        <Route
+          path="/settings"
+          element={
+            <AuthorizedRoute>
+              <SettingsPage />
+            </AuthorizedRoute>
+          }
+        />
+        <Route index element={<IndexPage />} />
         {/* For backwards compatibility */}
-        <Route
-          path="/photos"
-          exact
-          render={() => <Redirect to="/timeline" />}
-        />
-        <Route
-          render={() => (
-            <div>{t('routes.page_not_found', 'Page not found')}</div>
-          )}
-        />
-      </Switch>
+        <Route path="/photos" element={<Navigate to="/timeline" />} />
+        <Route element={<NotFoundPage t={t} />} />
+      </RouterRoutes>
     </React.Suspense>
   )
+}
+
+const IndexPage = () => {
+  const token = authToken()
+
+  const dest = token ? '/timeline' : '/login'
+
+  return <Navigate to={dest} />
+}
+
+export const NotFoundPage = ({ t }: { t: TFunction }) => {
+  return <div>{t('routes.page_not_found', 'Page not found')}</div>
+}
+
+const LogoutPage = ({ navigate }: { navigate: NavigateFunction }) => {
+  clearTokenCookie()
+  navigate('/')
+  return null
 }
 
 export default Routes
